refactor of header files
